move level action sequences into own folder
