+++ src/js/redux/actionSequences/level/checkTile.js
@@ -1,4 +1,4 @@
-import * as _ from '../allActions.js'
+import * as _ from '../../allActions.js'
 
 export default (tile) => (dispatch, getState) => {
   const { level: { isDragging }, updating } = getState()
