+++ src/js/redux/actionSequences/level/setUpBoard.js
@@ -1,6 +1,6 @@
 import Promise from 'bluebird'
-import { batch } from '../_thunkHelpers.js'
-import * as _ from '../allActions.js'
+import { batch } from '../../_thunkHelpers.js'
+import * as _ from '../../allActions.js'
 
 export default (goal, probabilities, boardSize = 8) => (dispatch) => {
   return Promise
